workspace/pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.5",
    "discord-py>=2.5.2",
    "dotenv>=0.9.9",
    "fastapi>=0.115.14",
    "freezegun>=1.5.2",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-mock>=3.14.1",
    "supabase>=2.16.0",
    "uvicorn>=0.35.0",
]
